requirements.txt
click==8.1.3
python-dotenv==1.0.0
Requests==2.31.0
setuptools==65.5.1
simple_term_menu==1.6.1
